[obey-campaign:8deed8b4] fix(switch): hop-back must not require a local registry
Self-review of the gesture found it sitting behind the 'no campaigns registered'
guard. Hop-back resolves the campaign on the ORIGIN machine's registry, so a
local one is not a precondition, and the guard's message would have sent an
operator to 'camp init' for a completely unrelated reason.

Hoisted the check above the registry load. It now also runs before the machines
file is read, which is correct for the same reason: the gesture reads only the
payload.

Also splits the selector and scope helpers into switch_selector.go, keeping
switch.go under the 500-line ceiling after the hoist (508 to 469). Pure move,
no behavior change.
